requirements.txt
# FIBO Omni-Director Pro - Streamlit Cloud Requirements
# Combines frontend and essential backend dependencies

# Frontend Requirements
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
pillow>=10.0.0

# Backend Core (Essential for Streamlit Cloud)
fastapi>=0.104.0
pydantic>=2.4.0
sqlalchemy>=2.0.0
sqlite3

# Image Processing & Brand Protection
pillow>=10.0.0
opencv-python-headless>=4.8.0

# API Clients
openai>=1.3.0
httpx>=0.25.0

# Data Processing
python-multipart>=0.0.6
python-dateutil>=2.8.0
uuid>=1.30

# Environment
python-dotenv>=1.0.0